some cnhgs

+++ tasks/stl/sharapov_g_sobel/include/ssobel_stl.hpp
@@ -31,7 +31,7 @@ class SSobelStl : public ppc::core::Task {
   static std::vector<GrayScale> SobelOperatorStl(const std::vector<GrayScale>& grayImage, size_t width, size_t height);
 
  private:
-  size_t imgWidth{}, imgHeight{}, imgSize{};
+  int imgWidth{}, imgHeight{}, imgSize{};
   std::vector<RGB> colored_img{};
   std::vector<GrayScale> grayscale_img{};
   std::vector<GrayScale> result{};
